tests/wpt/tests/websockets/closing-handshake/004.html

<!doctype html>
<title>WebSockets: data after closing handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.sub.js></script>
<meta name="variant" content="?default">
<meta name="variant" content="?wss">
<meta name="variant" content="?wpt_flags=h2">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo_close_data');
  ws.onmessage = ws.onerror = t.unreached_func();

  ws.onopen = t.step_func(function(e) {
    ws.send('Goodbye');
  })

  ws.onclose = t.step_func(function(e) {
    assert_equals(e.wasClean, true);
    ws.onclose = t.unreached_func();
    t.step_timeout(() => t.done(), 50);
  })
});
</script>
